requirements/base.txt
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-ckeditor==5.6.1
django-import-export==1.1.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-singleton-model==0.2.3
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
psycopg2-binary==2.7.6.1
pytz==2018.7
PyYAML==3.13
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
